+++ src/compiler/ast/src/lib.rs
@@ -144,6 +144,15 @@ pub enum Number {
     Float(f64),
 }
 
+impl std::fmt::Display for Number {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Number::Int(i) => write!(f, "{i}"),
+            Number::Float(fl) => write!(f, "{fl}"),
+        }
+    }
+}
+
 #[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
 pub enum Validator<'src> {
     // Numeric validators